Add NotFound component and move router logic to it's own file
