+++ packages/pxweb2/public/locales/ar/translation.json
@@ -9,6 +9,16 @@
       "title": "PxWeb 2.0 [Main]",
       "logo": "بيكس ويب 2.0"
     },
+    "not_found": {
+      "table_not_found": {
+        "title": "لم يتم العثور على الجدول",
+        "description": "الجدول الذي تبحث عنه غير موجود. يرجى التحقق من عنوان URL أو العودة إلى الصفحة الرئيسية."
+      },
+      "unsupported_language": {
+        "title": "لغة غير مدعومة",
+        "description": "اللغة المحددة غير مدعومة."
+      }
+    },
     "skip_to_main": "انتقل إلى المحتوى الرئيسي",
     "footer": {
       "contact": {
